guidance comments.
I said "easy and stress free". I had something easier in mind, but give this a shot :)

It' won't compile as it stands. take your time, fix it as you go. 

This exercise will be about "leveraging the power of arrays to turn complicated management of many, overly simple things" into "very simple management of simple, but not overly simple, bits of code.